+++ ngClient/_system/js/app/controllers/content/BasicsController.js
@@ -11,115 +11,15 @@
         logger = logger.forSource(controllerId);
 
         var vm = this;
-        vm.existingModelConfig = {};
-        vm.newModelConfig = {};
+        vm.existingModelConfig = { userName: 'sample', resourcePoolKey: 'Basics-Existing-Model' };
+        vm.newModelConfig = { userName: 'sample', resourcePoolKey: 'Basics-New-Model' };
 
         // Listen resource pool updated event
         $scope.$on('resourcePoolEditor_elementMultiplierIncreased', updateOppositeResourcePool);
         $scope.$on('resourcePoolEditor_elementMultiplierDecreased', updateOppositeResourcePool);
         $scope.$on('resourcePoolEditor_elementMultiplierReset', updateOppositeResourcePool);
 
-        _init();
-
         /*** Implementations ***/
-
-        function _init() {
-
-            dataContext.getCurrentUser()
-                .then(function (currentUser) {
-                    vm.existingModelConfig = { userName: currentUser.UserName, resourcePoolKey: 'Basics-Existing-Model' };
-                    vm.newModelConfig = { userName: currentUser.UserName, resourcePoolKey: 'Basics-New-Model' };
-
-                    resourcePoolFactory.getResourcePoolExpanded(vm.existingModelConfig)
-                        .then(function (resourcePool) {
-                            if (resourcePool === null) {
-                                getBasicsSample()
-                                    .then(function (resourcePool) {
-                                        dataContext.createEntitySuppressAuthValidation(true);
-
-                                        resourcePool.Name = 'Basics - Existing Model';
-                                        resourcePool.Key = vm.existingModelConfig.resourcePoolKey;
-                                        resourcePool.UserResourcePoolSet[0].entityAspect.setDeleted(); // Remove resource pool rate
-                                        resourcePool._init(true);
-
-                                        dataContext.createEntitySuppressAuthValidation(false);
-                                    });
-                            }
-                        });
-
-                    resourcePoolFactory.getResourcePoolExpanded(vm.newModelConfig)
-                        .then(function (resourcePool) {
-                            if (resourcePool === null) {
-                                getBasicsSample()
-                                    .then(function (resourcePool) {
-                                        dataContext.createEntitySuppressAuthValidation(true);
-
-                                        resourcePool.Name = 'Basics - New Model';
-                                        resourcePool.Key = vm.newModelConfig.resourcePoolKey;
-
-                                        // Employee Satisfaction field (index)
-                                        var employeeSatisfactionField = resourcePoolFactory.createElementField({
-                                            Element: resourcePool.mainElement(),
-                                            Name: 'Employee Satisfaction',
-                                            DataType: 4,
-                                            UseFixedValue: false,
-                                            IndexEnabled: true,
-                                            IndexCalculationType: 1,
-                                            IndexSortType: 1,
-                                            SortOrder: 2
-                                        });
-
-                                        employeeSatisfactionField.ElementCellSet.forEach(function (elementCell) {
-                                            var userElementCell = {
-                                                ElementCell: elementCell,
-                                                DecimalValue: Math.floor((Math.random() * 100) + 1)
-                                            };
-
-                                            dataContext.createEntity('UserElementCell', userElementCell);
-                                        });
-
-                                        resourcePool._init(true);
-
-                                        dataContext.createEntitySuppressAuthValidation(false);
-                                    });
-                            }
-                        });
-                });
-        }
-
-        function getBasicsSample() {
-
-            dataContext.createEntitySuppressAuthValidation(true);
-
-            return resourcePoolFactory.createResourcePoolDirectIncomeAndMultiplier()
-                .then(function (resourcePool) {
-                    dataContext.createEntitySuppressAuthValidation(true);
-
-                    resourcePool.InitialValue = 0;
-                    resourcePool.isTemp = true;
-
-                    var mainElement = resourcePool.mainElement();
-                    mainElement.Name = 'Organization';
-
-                    mainElement.ElementItemSet[0].Name = 'Alpha';
-                    mainElement.ElementItemSet[1].Name = 'Beta';
-                    resourcePoolFactory.createElementItem({
-                        Element: mainElement,
-                        Name: 'Charlie'
-                    });
-                    resourcePoolFactory.createElementItem({
-                        Element: mainElement,
-                        Name: 'Delta'
-                    });
-
-                    dataContext.createEntitySuppressAuthValidation(false);
-
-                    return resourcePool;
-                })
-                .finally(function () {
-                    dataContext.createEntitySuppressAuthValidation(false);
-                });
-        }
 
         function updateOppositeResourcePool(event, element) {
 
@@ -135,6 +35,7 @@
             if (oppositeKey !== null) {
                 resourcePoolFactory.getResourcePoolExpanded(oppositeKey)
                 .then(function (resourcePool) {
+
                     switch (event.name) {
                         case 'resourcePoolEditor_elementMultiplierIncreased': {
                             dataContext.updateElementMultiplier(resourcePool.mainElement(), 'increase');
@@ -149,6 +50,8 @@
                             break;
                         }
                     }
+
+                    dataContext.saveChanges(1500);
                 });
             }
         }