+++ ngClient/_system/js/app/entities/User.js
@@ -43,9 +43,6 @@
             //self.Logins = [];
             //self.Roles = [];
 
-            // Client-side
-            self.isEditing = false;
-
             // Functions
             self.isAuthenticated = isAuthenticated;