+++ ngClient/_system/js/app/entities/UserElementCell.js
@@ -42,7 +42,6 @@
             self.backingFields = {
                 _DecimalValue: null
             };
-            self.isEditing = false;
         }
     }
 })();